+++ packages/components/panel/src/panel.stories.ts
@@ -12,10 +12,12 @@ import {
   faTrash,
   faUnlock
 } from '@fortawesome/pro-regular-svg-icons';
+import { announce } from '@sl-design-system/announcer';
 import '@sl-design-system/button/register.js';
 import { Icon } from '@sl-design-system/icon';
+import { SlToggleEvent } from '@sl-design-system/shared/events.js';
 import { type Meta, type StoryObj } from '@storybook/web-components-vite';
-import { type TemplateResult, html } from 'lit';
+import { LitElement, type TemplateResult, html } from 'lit';
 import '../register.js';
 import { type Panel, type PanelDensity, PanelElevation, type TogglePlacement } from './panel.js';
 
@@ -225,6 +227,54 @@ export const OverflowActions: Story = {
   }
 };
 
+export const ActionsWithTooltips: Story = {
+  args: {
+    ...Basic.args,
+    actions: () => html`
+      <sl-button aria-labelledby="edit-tooltip" fill="ghost" slot="actions">
+        <sl-icon name="far-pen"></sl-icon>
+      </sl-button>
+      <sl-tooltip id="edit-tooltip">Edit</sl-tooltip>
+      <sl-button aria-labelledby="copy-tooltip" fill="ghost" slot="actions">
+        <sl-icon name="far-copy"></sl-icon>
+      </sl-button>
+      <sl-tooltip id="copy-tooltip">Copy</sl-tooltip>
+      <sl-button aria-labelledby="paste-tooltip" fill="ghost" slot="actions">
+        <sl-icon name="far-paste"></sl-icon>
+      </sl-button>
+      <sl-tooltip id="paste-tooltip">Paste</sl-tooltip>
+      <sl-button aria-labelledby="share-tooltip" fill="ghost" slot="actions">
+        <sl-icon name="far-share"></sl-icon>
+      </sl-button>
+      <sl-tooltip id="share-tooltip">Paste</sl-tooltip>
+    `,
+    content: () => 'This panel contains action buttons with tooltips.'
+  }
+};
+
+export const FixedInlineSize: Story = {
+  args: {
+    ...Basic.args,
+    actions: () => html`
+      <sl-button fill="ghost" slot="actions" aria-label="Copy">
+        <sl-icon name="far-copy"></sl-icon>
+      </sl-button>
+      <sl-button fill="ghost" slot="actions" aria-label="Book">
+        <sl-icon name="far-book"> </sl-icon>
+      </sl-button>
+    `,
+    content: () => html`
+      <style>
+        sl-panel {
+          inline-size: 350px;
+        }
+      </style>
+
+      This is a panel with a fixed inline size of 350px.
+    `
+  }
+};
+
 export const WithoutActions: Story = {
   args: {
     ...OverflowHeading.args,
@@ -325,6 +375,56 @@ export const NoHeader: Story = {
         <img alt="city" src="https://images.unsplash.com/photo-1586622992874-27d98f198139" />
       </div>
     `
+  }
+};
+
+export const ToggleExternally: Story = {
+  render: () => {
+    try {
+      customElements.define(
+        'panel-toggle-example',
+        class extends LitElement {
+          collapsed = false;
+
+          override render(): TemplateResult {
+            return html`
+              <h2>We use the announcer to inform the user, when the panel is opened/closed externally.</h2>
+              <sl-button @click=${this.togglePanel}>Toggle panel</sl-button>
+              <p>State: ${this.collapsed ? 'closed' : 'open'}</p>
+              <sl-panel
+                @sl-toggle=${(e: SlToggleEvent) => this.onToggle(e)}
+                collapsible
+                heading="Discovering Dinosaurs 🦕"
+              >
+                Embark on a thrilling journey back in time to the age of dinosaurs, where colossal creatures roamed the
+                Earth 🌎 and ancient landscapes teemed with life. This prehistoric adventure invites you to explore a
+                world untouched by modern civilization, filled with towering ferns 🌿, volcanic mountains, and the
+                thunderous footsteps of giants like the Tyrannosaurus rex 🦖 and the Triceratops. Whether you're
+                navigating dense jungles or witnessing dramatic encounters between predator and prey, every moment is
+                packed with excitement and discovery. Get ready to uncover the mysteries of a lost era and experience
+                the awe-inspiring power of nature in its most primal form.
+              </sl-panel>
+            `;
+          }
+
+          togglePanel() {
+            this.renderRoot.querySelector('sl-panel')?.toggle();
+            announce(`Panel ${this.collapsed ? 'collapsing' : 'expanding'}`);
+            this.requestUpdate();
+          }
+
+          //make sure that the state of the panel is updated in the current component when it's changed in the SLDS panel component
+          onToggle(event: SlToggleEvent) {
+            this.collapsed = event.detail as boolean;
+            this.requestUpdate();
+          }
+        }
+      );
+    } catch {
+      /* empty */
+    }
+
+    return html`<panel-toggle-example></panel-toggle-example>`;
   }
 };